Move the SDK to devDependencies and peerDependencies
The Stakeholder SDK is an extension library of the SDK.
The user of Stakeholder SDK should install the user's SDK with the
Stakeholder SDK. Peer dependency works for the case. When using
peer dependency, NPM or Yarn does not install the SDK when installing
the Stakeholder SDK. However, they complain that the peer dependent
library is not installed.

Also, we need the SDK's type to build Typescript code. If we use the
dev dependency, we use the SDK code only for the development environment.